[Perf] (drawing|buffer): Optimize draw call overhead and state query performance
- Add g_current_program_needs_sampler_emulation flag to avoid per-draw map lookups

- Use thread_local vector with reserve for glDrawElementsBaseVertex temp buffer reuse

- Cache current VAO's IBO to eliminate vector lookup in find_bound_buffer

- Add binding_query_to_index for direct array indexing in find_bound_buffer

- Pre-store ignore_error_level at shader/program creation to avoid repeated global checks

+++ MobileGlues-cpp/gl/shader.cpp
@@ -126,7 +126,7 @@ void glGetShaderiv(GLuint shader, GLenum pname, GLint* params) {
         if (it != g_shaderInfos.end()) {
             ignore_level = it->second.ignore_error_level;
         }
-        if (ignore_level >= IgnoreErrorLevel::Partial) {
+        if (ignore_level >= static_cast<int>(IgnoreErrorLevel::Partial)) {
             GLchar infoLog[512];
             GLES.glGetShaderInfoLog(shader, 512, nullptr, infoLog);
             LOG_W_FORCE("Shader %d compilation failed: \n%s", shader, infoLog)
@@ -149,7 +149,7 @@ GLuint glCreateShader(GLenum shaderType) {
     GLuint shader = GLES.glCreateShader(shaderType);
 
     if (shader != 0) {
-        g_shaderInfos[shader].ignore_error_level = global_settings.ignore_error;
+        g_shaderInfos[shader].ignore_error_level = static_cast<int>(global_settings.ignore_error);
         if (hardware->emulate_texture_buffer) {
             shader_map_is_sampler_buffer_emulated[shader] = false;
         }
